Remove processorToPipe from interop.flow package object
